*BIG* UI refactor for clarity Renamed classes to have shorter names Split out UI utils into `BuildConfigEditorStyle` TODO: Cache all of the UI states, strings, LINQ queries TODO: Show the selected build platforms so users don't have to drill down to see a summary TODO: Fix warning icon scaling in the foldouts
