+++ workers/unity/Assets/Playground/Config/BuildConfiguration.asset
@@ -18,105 +18,68 @@ MonoBehaviour:
     - {fileID: 102900000, guid: 3f6e3d421576ca644b07380deacd35bb, type: 3}
     LocalBuildConfig:
       BuildTargets:
-      - Options: 1
+      - Options: 0
+        Target: 19
         Enabled: 1
       - Options: 0
+        Target: 13
         Enabled: 0
       - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
+        Target: 9
         Enabled: 0
     CloudBuildConfig:
       BuildTargets:
       - Options: 0
+        Target: 5
         Enabled: 0
       - Options: 0
+        Target: 19
+        Enabled: 1
+      - Options: 16384
+        Target: 24
         Enabled: 0
       - Options: 0
+        Target: 2
         Enabled: 1
       - Options: 0
+        Target: 13
         Enabled: 0
       - Options: 0
-        Enabled: 1
-      - Options: 0
-        Enabled: 0
-      - Options: 0
+        Target: 9
         Enabled: 0
   - WorkerType: UnityGameLogic
     ScenesForWorker:
     - {fileID: 102900000, guid: f5fda56aad5d91d408ff123b26ab6ade, type: 3}
     LocalBuildConfig:
       BuildTargets:
-      - Options: 540673
+      - Options: 0
+        Target: 19
         Enabled: 1
-      - Options: 1
+      - Options: 0
+        Target: 13
         Enabled: 0
-      - Options: 1
-        Enabled: 0
-      - Options: 1
-        Enabled: 0
-      - Options: 1
-        Enabled: 0
-      - Options: 1
-        Enabled: 0
-      - Options: 1
+      - Options: 0
+        Target: 9
         Enabled: 0
     CloudBuildConfig:
       BuildTargets:
       - Options: 0
+        Target: 5
         Enabled: 0
       - Options: 0
-        Enabled: 0
-      - Options: 0
+        Target: 19
         Enabled: 0
       - Options: 16384
+        Target: 24
         Enabled: 1
       - Options: 0
+        Target: 2
         Enabled: 0
       - Options: 0
+        Target: 13
         Enabled: 0
       - Options: 0
-        Enabled: 0
-  - WorkerType: AndroidClient
-    ScenesForWorker:
-    - {fileID: 102900000, guid: 82ccd0fd96a2ba24f88d0c7d01592f2e, type: 3}
-    LocalBuildConfig:
-      BuildTargets:
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 1
-        Enabled: 1
-      - Options: 0
-        Enabled: 0
-    CloudBuildConfig:
-      BuildTargets:
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 1
-      - Options: 0
+        Target: 9
         Enabled: 0
   - WorkerType: iOSClient
     ScenesForWorker:
@@ -124,33 +87,66 @@ MonoBehaviour:
     LocalBuildConfig:
       BuildTargets:
       - Options: 0
+        Target: 19
         Enabled: 0
       - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
+        Target: 13
         Enabled: 0
       - Options: 1
+        Target: 9
         Enabled: 1
     CloudBuildConfig:
       BuildTargets:
       - Options: 0
+        Target: 5
         Enabled: 0
       - Options: 0
+        Target: 19
+        Enabled: 0
+      - Options: 16384
+        Target: 24
         Enabled: 0
       - Options: 0
+        Target: 2
         Enabled: 0
       - Options: 0
+        Target: 13
         Enabled: 0
       - Options: 0
-        Enabled: 0
-      - Options: 0
-        Enabled: 0
-      - Options: 0
+        Target: 9
         Enabled: 1
+  - WorkerType: AndroidClient
+    ScenesForWorker:
+    - {fileID: 102900000, guid: 82ccd0fd96a2ba24f88d0c7d01592f2e, type: 3}
+    LocalBuildConfig:
+      BuildTargets:
+      - Options: 0
+        Target: 19
+        Enabled: 0
+      - Options: 0
+        Target: 13
+        Enabled: 1
+      - Options: 0
+        Target: 9
+        Enabled: 0
+    CloudBuildConfig:
+      BuildTargets:
+      - Options: 0
+        Target: 5
+        Enabled: 0
+      - Options: 0
+        Target: 19
+        Enabled: 0
+      - Options: 16384
+        Target: 24
+        Enabled: 0
+      - Options: 0
+        Target: 2
+        Enabled: 0
+      - Options: 0
+        Target: 13
+        Enabled: 1
+      - Options: 0
+        Target: 9
+        Enabled: 0
   isInitialised: 1
