+++ workers/unity/Assets/Playground/Config/BuildConfiguration.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: de6dc0c79c70a0f4491a1507aef63a4a
+guid: 32794e325bf25864cbb46d1f39f3458c
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
