+++ workers/unity/Packages/com.improbable.gdk.buildsystem/Configuration/BuildConfigEditorStyle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41fe4727d2a59f34596b510fc07c4bb8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
