+++ workers/unity/Packages/com.improbable.gdk.buildsystem/Configuration/BuildTargetConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08ccb7d4f2ab1374a8d5678ff5cc58a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
