+++ workers/unity/Packages/com.improbable.gdk.tools/.CodeGenerator/GdkCodeGenerator/Templates/UnityWorkerMenuGenerator.tt
@@ -63,40 +63,11 @@ var workerType = workerTypeList[i];
             Debug.Log("Clean completed");
         }
 
-        private static bool CheckWorkersCanBuildForEnvironment(BuildEnvironment environment, string[] workerTypes)
-        {
-            var canInitiateBuild = true;
-            var spatialOSBuildConfiguration = SpatialOSBuildConfiguration.GetInstance();
-
-            foreach (var workerType in workerTypes)
-            {
-                var buildTargetsForWorker = WorkerBuilder.GetBuildTargetsForWorkerForEnvironment(workerType, environment);
-                var buildTargetsMissingBuildSupport = BuildSupportChecker.GetBuildTargetsMissingBuildSupport(buildTargetsForWorker);
-
-                if (buildTargetsMissingBuildSupport.Length > 0)
-                {
-                    canInitiateBuild = false;
-
-                    Debug.LogError(BuildSupportChecker.ConstructMissingSupportMessage(workerType, environment, buildTargetsMissingBuildSupport),
-                        spatialOSBuildConfiguration);
-                }
-            }
-
-            return canInitiateBuild;
-        }
-
         private static void MenuBuild(BuildEnvironment environment, params string[] workerTypes)
         {
             // Delaying build by a frame to ensure the editor has re-rendered the UI to avoid odd glitches.
             EditorApplication.delayCall += () =>
             {
-                if (!CheckWorkersCanBuildForEnvironment(environment, workerTypes))
-                {
-                    DisplayBuildFailureDialog();
-
-                    return;
-                }
-
                 try
                 {
                     LocalLaunch.BuildConfig();
@@ -122,25 +93,6 @@ var workerType = workerTypeList[i];
             EditorUtility.DisplayDialog("Build Failed",
                 "Build failed. Please see the Unity Console Window for information.",
                 "OK");
-        }
-
-        [MenuItem(EditorConfig.ParentMenu + "/Check build support", false, EditorConfig.MenuOffset + <#= workerTypeList.Count + 1 #>)]
-        private static void CheckBuildSupport()
-        {
-            bool checksPassed = CheckWorkersCanBuildForEnvironment(BuildEnvironment.Local, AllWorkers);
-
-            if (!CheckWorkersCanBuildForEnvironment(BuildEnvironment.Cloud, AllWorkers))
-            {
-                checksPassed = false;
-            }
-
-            EditorUtility.DisplayDialog("Build Support Check",
-                checksPassed
-                    ? "Build support check passed. Your Unity Editor installation has all necessary components."
-                    : "Build support check failed. Please see the Unity Console Window for information.",
-                "OK");
-
-            Debug.Log("Build support check completed.");
         }
     }
 }
